+++ .github/workflows/go.yml
@@ -16,9 +16,9 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        go: ["1.19.x", "1.20.x"]
+        go: ["1.20.x", "1.21.x"]
         include:
-        - go: 1.20.x
+        - go: 1.21.x
           latest: true
 
     steps:
