+++ go.mod
@@ -1,6 +1,6 @@
 module go.uber.org/multierr
 
-go 1.19
+go 1.20
 
 require github.com/stretchr/testify v1.7.0
