+++ tools/go.mod
@@ -1,6 +1,6 @@
 module go.uber.org/multierr/tools
 
-go 1.18
+go 1.20
 
 require (
 	golang.org/x/lint v0.0.0-20210508222113-6edffad5e616
